Don't require a staging repository
A staging/side-band repository is nice to have in principle, but the
ability to cherry-pick & merge directly out of phabricator makes it not
a hard requirement.  Also depending on your workflow, it might required
extra effort to setup (doubly so for a public space) just to please
git-phab.

With these changes, the feature are still there but only enable if
the remote is configured.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Differential Revision: https://phabricator.freedesktop.org/D832