Don't ask for new task creating by default
Creating a new task when attaching patches is a somewhat odd workflow as
it really means the task is created after big part of the work has been
done.  A more typical workflow is either to create a task ahead of time
and attach patches later or for simple patches (that don't really
require a task) to only submit the patch without creating a task.

Change things such that by default git phab attach doesn't ask for task
creation, but allow that behaviour to be enabled by using --task without
an argument.

Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>
Differential Revision: https://phabricator.freedesktop.org/D835