git-phab: Increase result limit on repository.query method
Some Phabricator installations have more than 100 repositories; and 100
is the default results limit on repository.query, which can cause
git-phab to incorrectly not find the repository the user is trying to
submit to.

Ideally, we would be able to use the ‘remoteURIs’ parameter of
repository.query to instantly select the repository we care about (as we
essentially currently implement that search locally), but its semantics
are unclear.

Signed-off-by: Philip Withnall <philip.withnall@collabora.co.uk>
Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1250