git-phab: Add --never-apply-patches argument to `arc lint`
The linting infrastructure in git-phab assumes that the linter is not
interactive; if it is (as `arc lint` can sometimes be, for example when
running its clang-format plugin), everything will seemingly hang, as
git-phab is waiting for `arc lint` to complete before it will print the
output from it, and `arc lint` is waiting for user input after printing
a prompt which the user has not seen.

Fix that by automatically answering negatively to all `arc lint`
questions, which are typically of the form ‘do you want to make this
automatic indentation change?’. The questions should then be printed
afterwards.

Signed-off-by: Philip Withnall <philip.withnall@collabora.co.uk>
Differential Revision: https://phabricator.freedesktop.org/D1251