git-phab: Fix offsets in manually constructed git diffs
The new and old offsets of the start of the file in diffs which are sent
to Phabricator must only be zero if the file has been deleted or added,
respectively. This was previously backwards in one of the cases,
resulting in invalid raw diffs being downloaded from Phabricator, and
hence the round-trip `git phab attach; git phab merge` not working.

Signed-off-by: Philip Withnall <philip.withnall@collabora.co.uk>
Differential Revision: https://phabricator.freedesktop.org/D1254